server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
structlog>=24.1.0
python-dotenv>=1.0.0
motor>=3.3.0
beanie>=1.25.0
firebase-admin>=6.4.0
google-genai>=1.0.0
sse-starlette>=2.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
ruff>=0.2.0
mongomock-motor>=0.0.30
